Changes to allow manual locking/unlocking of a module
There are custom file transfer methods that may be needed.  This added a htc class and main pipeline code  that will allow for locking the module (not allow a module to execute) while the script is running.  This may be advantageous if you need to use a file transfer protocol not built into the serpent, in which you do not want to operate on a partial file.  This is controlled via a config .json key  "lock_remote": "<modulename>",